FEAT: Pre and Post generation automation.
Some improvements were made.

Pre prompting hook is added to check if git is installed into the
system. Post Project Generation hook is also added to sync the generated
template with desired repo for modules fetching and to modify the config
file to set a dynamic SHA-1 admin_passwd into it.